Fix resolving page name for link
Some links (created by modules) do not set 'page' property. This leads
to PHP warnings later.

Of course, best way to handle this is to fix the modules that integrates
with sitemap. Unfortunately that is not possible, so we have at least
update the sitemap module to survive this situation